+++ resources/views/activity/overview.blade.php
@@ -164,7 +164,7 @@
 
 @push('scripts')
     @if ($hasData)
-        <script type="module" src="{{ route('elastic-audit.assets', ['asset' => $elasticAuditAssets['resources/js/chart.js']['file']]) }}"></script>
+        <script type="module" src="{{ route('elastic-audit.assets', ['asset' => $elasticAuditAssets['resources/js/chart.js']['file']], false) }}"></script>
         <script type="module">
             (function () {
                 const tz       = @json($timezone);